+++ src/main/data/duke.txt
@@ -1,0 +1,3 @@
+D | 0 | test | now
+T | 0 | test
+T | 0 | 1
