--- src/main/java/command/Commands.java
@@ -1,7 +1,0 @@
-package command;
-
-public enum Commands {
-    LIST, MARK, UNMARK, DEADLINE, TODO, EVENT, DELETE
-}
-
-
