+++ src/main/java/duke/Duke.java
@@ -1,15 +1,23 @@
-import java.util.Scanner;
-import ui.Ui;
-import storage.Storage;
-import taskList.TaskList;
-import parser.Parser;
+package duke;
 
+import java.util.Scanner;
+import duke.ui.Ui;
+import duke.storage.Storage;
+import duke.taskList.TaskList;
+import duke.parser.Parser;
+
+/**
+ * A chatbot inspired by the Java Mascot: duke.Duke.
+ */
 public class Duke {
 
     private Ui ui;
     private TaskList tasks;
     private Storage storage;
 
+    /**
+     * Constructor for duke.Duke.
+     */
     public Duke() {
         this.ui = new Ui();
         this.storage = new Storage();
@@ -17,14 +25,18 @@ public class Duke {
         storage.load(this.tasks);
     }
 
+    /**
+     * This runs the duke programs main chat functionality,
+     * the while loop continues until user types bye.
+     */
     public void run() {
         Scanner sc = new Scanner(System.in);
         Parser parser = new Parser();
 
-        while(!false) {
+        while(true) {
             String command = sc.nextLine();
-            int i = parser.parseCommand(command, tasks);
-            if (i == 1) { break; }
+            boolean isBye = parser.parseCommand(command, tasks);
+            if (isBye) { break; }
         }
     }
