+++ src/main/java/duke/command/Commands.java
@@ -1,0 +1,10 @@
+package duke.command;
+
+/**
+ * An enum of commands.
+ */
+public enum Commands {
+    LIST, MARK, UNMARK, DEADLINE, TODO, EVENT, DELETE, FIND
+}
+
+
