+++ src/main/java/duke/parser/Parser.java
@@ -1,20 +1,17 @@
-package parser;
+package duke.parser;
 
-import java.util.ArrayList;
-import java.util.Scanner;
 import java.util.regex.Pattern;
-import command.Commands;
-import dukeException.DukeException;
-import task.Deadline;
-import task.Event;
-import task.Task;
-import task.Todo;
-import taskList.TaskList;
+import duke.command.Commands;
+import duke.dukeException.DukeException;
+import duke.taskList.TaskList;
 
 /**
- * deals with making sense of the user command
+ * deals with making sense of the user duke.command.
  */
 public class Parser {
+    /**
+     * The constructor for the Parser class.
+     */
     public Parser() {
 
     }
@@ -25,42 +22,51 @@ public class Parser {
     Pattern todoRegex = Pattern.compile("^" + Commands.TODO, Pattern.CASE_INSENSITIVE);
     Pattern eventRegex = Pattern.compile("^" + Commands.EVENT, Pattern.CASE_INSENSITIVE);
     Pattern deleteRegex = Pattern.compile("^" + Commands.DELETE, Pattern.CASE_INSENSITIVE);
+    Pattern findRegex = Pattern.compile("^" + Commands.FIND, Pattern.CASE_INSENSITIVE);
 
-    public int parseCommand(String command, TaskList taskList) {
-        ArrayList<Task> list = taskList.getList();
+
+    /**
+     * This will be responsible for parsing user duke.command
+     * and the duke.command will be acted out on the TaskList input.
+     * @param command User input to be parsed by duke.parser.
+     * @param taskList The Tasklist that the duke.command would be used on.
+     * @return boolean Returns true if user inputs "bye".
+     */
+    public boolean parseCommand(String command, TaskList taskList) {
         if (command.equals("bye")) {
-            return 1;
+            return true; // when returning 0, the application closes
         } else if (listRegex.matcher(command).find()) {
-            Task.printList(list);
-            return 0;
+            taskList.printList();
+            return false;
         } else if (markRegex.matcher(command).find()) {
             int curr = Integer.parseInt(command.substring(5)) - 1;
-            Task currTask = list.get(curr);
-            currTask.markDone();
-            return 0;
+            taskList.markCurrentTaskDone(curr);
+            return false;
         } else if (unmarkRegex.matcher(command).find()) {
             int curr = Integer.parseInt(command.substring(7)) - 1;
-            Task currTask = list.get(curr);
-            currTask.markUndone();
-            return 0;
+            taskList.markCurrentTaskUndone(curr);
+            return false;
         } else if (deadlineRegex.matcher(command).find()) {
-            Deadline.addDeadline(list, command);
-            return 0;
+            taskList.addDeadline(command);
+            return false;
         } else if (todoRegex.matcher(command).find()) {
-            Todo.addTodo(list, command);
-            return 0;
+            taskList.addTodo(command);
+            return false;
         } else if (eventRegex.matcher(command).find()) {
-            Event.addEvent(list, command);
-            return 0;
+            taskList.addEvent(command);
+            return false;
         } else if (deleteRegex.matcher(command).find()) {
-            Task.deleteTask(list, command);
-            return 0;
+            taskList.deleteTask(command);
+            return false;
+        } else if (findRegex.matcher(command).find()) {
+            taskList.findTasks(command);
+            return false;
         } else {
             try {
                 throw new DukeException("Invalid Response");
             } catch (DukeException e) {
                 System.out.println("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
-                return 0;
+                return false;
             }
         }
     }
