--- src/main/java/dukeException/DukeException.java
@@ -1,15 +1,0 @@
-package dukeException;
-
-public class DukeException extends Exception {
-
-//    private String errMsg;
-    public DukeException(String errMsg)  {
-        super(errMsg);
-    }
-
-//    public String getErrorMessage() {
-//       return this.errMsg;
-//    }
-
-
-}
