+++ text-ui-test/ACTUAL.TXT
@@ -17,9 +17,3 @@ Got it, I've added this event.
 [E][ ] project meeting  (from: Mon 2pm to: 4pm)
 Now you have 3 tasks in the list.
 __________________________________________________
-__________________________________________________
-Here are the tasks in your list:
-1.[T][ ] borrow book
-2.[D][ ] return book  (by: Sunday)
-3.[E][ ] project meeting  (from: Mon 2pm to: 4pm)
-__________________________________________________
